dont operate on non-initialized browser

+++ dcumiddleware/urihelper.py
@@ -76,6 +76,7 @@ class URIHelper:
         :return:
         """
         data = (None, None)
+        browser = None
         try:
             browser = webdriver.PhantomJS()
             browser.set_page_load_timeout(10)
@@ -88,8 +89,9 @@ class URIHelper:
             self._logger.error("Error while taking snapshot and/or source code for %s: %s", url, str(e))
             return data
         finally:
-            browser.service.process.send_signal(signal.SIGTERM)
-            browser.quit()
+            if browser:
+                browser.service.process.send_signal(signal.SIGTERM)
+                browser.quit()
 
     def get_status(self, sourceDomainOrIp):
         """
